pyproject.toml
[project]
name = "find-my-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "geopy>=2.4.1",
    "pydantic>=2.11.5",
    "pyicloud>=1.0.0",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.3",
]
